requirements.txt
langchain>=1.2.10
langchain-openai>=1.1.9
langgraph>=1.0.8
langgraph-checkpoint-sqlite>=3.0.3
openai>=2.20.0
pre-commit>=4.5.1
tavily-python>=0.7.21
